pyproject.toml
[tool.poetry]
name = "tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow-providers-cncf-kubernetes = "^10.1.0"
altair = "5.0.0"
annotated-types = "0.7.0"
appnope = "0.1.4"
arrow = "1.3.0"
asttokens = "2.4.1"
attrs = "23.2.0"
autocommand = "2.2.2"
backports-tarfile = "1.1.1"
blinker = "1.8.1"
cachetools = "5.3.3"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cheroot = "10.0.1"
cherrypy = "18.9.0"
click = "8.1.7"
comm = "0.2.2"
contourpy = "^1.3.1"
cycler = "0.12.1"
debugpy = "1.8.1"
decorator = "5.1.1"
executing = "2.0.1"
fitbit = "0.3.1"
fonttools = "4.51.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
idna = "3.7"
importlib-metadata = "6.8.0"
inflect = "7.2.1"
ipykernel = "6.29.4"
ipython = "^8.31.0"
jaraco-collections = "5.0.1"
jaraco-context = "5.3.0"
jaraco-functools = "4.0.1"
jaraco-text = "3.12.0"
jedi = "0.19.1"
jinja2 = "3.1.3"
jsonschema = "4.22.0"
jsonschema-specifications = "2023.12.1"
jupyter-client = "8.6.1"
jupyter-core = "5.7.2"
kiwisolver = "1.4.5"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
matplotlib = "3.7.5"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
more-itertools = "10.2.0"
nest-asyncio = "1.6.0"
numpy = "1.24.4"
oauthlib = "3.2.2"
packaging = "24.0"
pandas = "2.0.3"
parso = "0.8.4"
pexpect = "4.9.0"
pillow = "10.3.0"
platformdirs = "4.2.1"
portend = "3.2.0"
prompt-toolkit = "3.0.43"
protobuf = "4.25.3"
psutil = "5.9.8"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pyarrow = "16.0.0"
pydantic = "2.7.1"
pydantic-settings = "2.2.1"
pydantic-core = "2.18.2"
pydeck = "0.9.0"
pygments = "2.17.2"
pyparsing = "3.1.2"
python-dateutil = "2.9.0.post0"
python-decouple = "3.8"
python-dotenv = "1.0.1"
pytz = "2024.1"
pyzmq = "26.0.3"
referencing = "0.35.1"
requests = "2.31.0"
requests-oauthlib = "2.0.0"
rich = "13.7.1"
rpds-py = "0.18.0"
seaborn = "0.13.2"
six = "1.16.0"
smmap = "5.0.1"
stack-data = "0.6.3"
streamlit = "1.4.0"
tempora = "5.5.1"
tenacity = "8.2.3"
toml = "0.10.2"
toolz = "0.12.1"
tornado = "6.4"
traitlets = "5.14.3"
typeguard = "4.2.1"
types-python-dateutil = "2.9.0.20240316"
typing-extensions = "4.11.0"
tzdata = "2024.1"
tzlocal = "5.2"
urllib3 = "2.2.1"
validators = "0.22.0"
wcwidth = "0.2.13"
zc-lockfile = "3.0.post1"
zipp = "3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
